Fixed #23 Return Duration from model not string
