+++ lib/models/speech_status.dart
