+++ lib/models/timeable.dart
