+++ lib/screens/timer/src/timer_buttons/src/button_properties.dart
