+++ lib/utilities/debate_events/src/lincoln_douglas.dart
