--- lib/utilities/utilities.dart
@@ -1,9 +1,0 @@
-library utilities;
-
-import 'dart:io';
-
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-part 'src/clear_timer.dart';
